+++ src/data/Assignment.kt
@@ -2,8 +2,10 @@ package data
 
 data class AssignmentCfg(
 	val workers: List<String> = emptyList(),
+	val capacity: List<Double> = emptyList(),
 	val tasks: List<String> = emptyList(),
 	val cost: List<List<Double>> = emptyList(),
+	val value: List<List<Double>> = emptyList(),
 	val teams: List<List<String>> = emptyList(),
 	val maxPerTeam: Int = 0,
 ) {
@@ -12,23 +14,27 @@ data class AssignmentCfg(
 	}
 
 	companion object {
-		fun new(name: String): AssignmentCfg? {
+		fun new(name: String, mustBeEqual: Boolean): AssignmentCfg? {
 			val data = load(name) ?: return null
 			val workers = data["workers"].toStringList()
 			val tasks = data["tasks"].toStringList()
-			if (tasks.size > workers.size) {
+			if (tasks.size > workers.size && mustBeEqual) {
 				println("Invalid assignment problem: more tasks than workers")
 				return null
 			}
 			val cost = mutableListOf<List<Double>>()
 			for(line in data["cost"].parseList()) {
 				val row = line.matrixRow(true)
-				val costRow = if (row.size < workers.size) {
+				val costRow = if (row.size < workers.size && mustBeEqual) {
 					row + List(workers.size - row.size) { 0.0 }
 				} else  {
 					row.subList(0, workers.size)
 				}
 				cost.add(costRow)
+			}
+			val value = mutableListOf<List<Double>>()
+			for(line in data["value"].parseList()) {
+				value.add(line.matrixRow(true))
 			}
 			val teams = mutableListOf<List<String>>()
 			for(line in data["teams"].parseList()) {
@@ -36,8 +42,10 @@ data class AssignmentCfg(
 			}
 			return AssignmentCfg(
 				workers = workers,
+				capacity = data["capacity"].toDoubleList(),
 				tasks = tasks,
 				cost = cost,
+				value = value,
 				teams = teams,
 				maxPerTeam = data["maxPerTeam"].parseInt(),
 			)
@@ -69,6 +77,35 @@ data class QuadraticAssignment(
 				count = data["count"].parseInt(),
 				distance = distance,
 				flow = flow,
+			)
+		}
+	}
+}
+
+data class Weapons(
+	val weapons: List<String> = emptyList(),
+	val count: List<Int> = emptyList(),
+	val targets: List<String> = emptyList(),
+	val value: List<Double> = emptyList(),
+	val chance: List<List<Double>> = emptyList(),
+) {
+	override fun toString(): String {
+		return "Weapons: $weapons\nCount: $count\nTargets: $targets\nValue: $value"
+	}
+
+	companion object {
+		fun new(name: String): Weapons? {
+			val data = load(name) ?: return null
+			val chance = mutableListOf<List<Double>>()
+			for(line in data["chance"].parseList()) {
+				chance.add(line.matrixRow(true))
+			}
+			return Weapons(
+				weapons = data["weapons"].toStringList(),
+				targets = data["targets"].toStringList(),
+				count = data["count"].toIntList(),
+				value = data["value"].toDoubleList(),
+				chance = chance,
 			)
 		}
 	}